refactor(fare): 요금을 계산하는 로직 분리
